deffered lib actually done
