+++ src/infrastructure/libs/deffered/defferedComponents/Await/Await.tsx
@@ -7,7 +7,6 @@ type PropsType = {
 	// call getData and add wait for response promise in maybe class store for this tasks
 	// and after promise resolved send serialized promoise response to client
 	// do all this things ONLY IF SERVER
-	//
 	getData: () => Promise<any>
 	// should be typed
 	children: (deffered: Promise<any>) => React.ReactNode;
