+++ src/infrastructure/libs/deffered/defferedComponents/context/useGetDefferedStore.tsx
@@ -4,5 +4,5 @@ import { DefferedStoreContext } from './context';
 export const useGetDefferedPromise = (name: string) => {
 	const store = useContext(DefferedStoreContext);
 
-	return store.get(name);
+	return store.getByActionName(name);
 };
