+++ src/infrastructure/libs/deffered/index.ts
@@ -1,0 +1,5 @@
+import { Await } from './defferedComponents/Await/Await';
+import { getDefferedStore } from './defferedStore/defferedStore';
+import { DefferedStoreContext } from './defferedComponents/context/context';
+
+export {Await, getDefferedStore, DefferedStoreContext};
